temp: myChallenge

+++ applications/challenge-application/src/main/java/com/pda/challengeapplication/challenges/repository/ChallengeDetailRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface ChallengeDetailRepository extends JpaRepository<ChallengeDetail, Integer> {
-    Optional<ChallengeDetail> findByChallengeId(Integer ChallengeId);
+    Optional<ChallengeDetail> findByChallengeId(long ChallengeId);
     ChallengeDetail save(ChallengeDetail challengeDetail);
 
     List<ChallengeDetail> findAll();
